+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/Images.ts
@@ -2,44 +2,30 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import * as THREE from "three";
-
 import { MultiMap } from "@foxglove/den/collection";
-import {
-  PinholeCameraModel,
-  decodeYUV,
-  decodeRGB8,
-  decodeRGBA8,
-  decodeBGRA8,
-  decodeBGR8,
-  decodeFloat1c,
-  decodeBayerRGGB8,
-  decodeBayerBGGR8,
-  decodeBayerGBRG8,
-  decodeBayerGRBG8,
-  decodeMono8,
-  decodeMono16,
-  decodeYUYV,
-} from "@foxglove/den/image";
+import { PinholeCameraModel } from "@foxglove/den/image";
 import Logger from "@foxglove/log";
 import { toNanoSec } from "@foxglove/rostime";
 import { CameraCalibration, CompressedImage, RawImage } from "@foxglove/schemas";
 import { SettingsTreeAction, SettingsTreeFields, Topic } from "@foxglove/studio";
-import { AnyImage } from "@foxglove/studio-base/panels/ThreeDeeRender/renderables/ImageTypes";
-import type { RosValue } from "@foxglove/studio-base/players/types";
+import { ImageRenderable } from "@foxglove/studio-base/panels/ThreeDeeRender/renderables/Images/ImageRenderable";
+import { AnyImage } from "@foxglove/studio-base/panels/ThreeDeeRender/renderables/Images/ImageTypes";
+import {
+  normalizeCompressedImage,
+  normalizeRawImage,
+  normalizeRosCompressedImage,
+  normalizeRosImage,
+} from "@foxglove/studio-base/panels/ThreeDeeRender/renderables/Images/imageNormalizers";
 
-import { cameraInfosEqual, normalizeCameraInfo, projectPixel } from "./projections";
+import { cameraInfosEqual, normalizeCameraInfo } from "./projections";
 import type { IRenderer } from "../IRenderer";
-import { BaseUserData, Renderable } from "../Renderable";
-import { PartialMessage, PartialMessageEvent, SceneExtension } from "../SceneExtension";
+import { PartialMessageEvent, SceneExtension } from "../SceneExtension";
 import { SettingsTreeEntry } from "../SettingsManager";
-import { stringToRgba } from "../color";
 import {
   CAMERA_CALIBRATION_DATATYPES,
   COMPRESSED_IMAGE_DATATYPES,
   RAW_IMAGE_DATATYPES,
 } from "../foxglove";
-import { normalizeByteArray, normalizeHeader, normalizeTime } from "../normalizeMessages";
 import {
   CameraInfo,
   Image as RosImage,
@@ -63,10 +49,8 @@ export type LayerSettingsImage = BaseSettings & {
 };
 
 const NO_CAMERA_INFO_ERR = "NoCameraInfo";
-const CREATE_BITMAP_ERR = "CreateBitmap";
 const CAMERA_MODEL = "CameraModel";
 
-const DEFAULT_IMAGE_WIDTH = 512;
 const DEFAULT_DISTANCE = 1;
 const DEFAULT_PLANAR_PROJECTION_FACTOR = 0;
 
@@ -78,31 +62,6 @@ const DEFAULT_SETTINGS: LayerSettingsImage = {
   planarProjectionFactor: DEFAULT_PLANAR_PROJECTION_FACTOR,
   color: "#ffffff",
 };
-
-export type ImageUserData = BaseUserData & {
-  topic: string;
-  settings: LayerSettingsImage;
-  cameraInfo: CameraInfo | undefined;
-  cameraModel: PinholeCameraModel | undefined;
-  image: AnyImage | undefined;
-  texture: THREE.Texture | undefined;
-  material: THREE.MeshBasicMaterial | undefined;
-  geometry: THREE.PlaneGeometry | undefined;
-  mesh: THREE.Mesh | undefined;
-};
-
-export class ImageRenderable extends Renderable<ImageUserData> {
-  public override dispose(): void {
-    this.userData.texture?.dispose();
-    this.userData.material?.dispose();
-    this.userData.geometry?.dispose();
-    super.dispose();
-  }
-
-  public override details(): Record<string, RosValue> {
-    return { image: this.userData.image, camera_info: this.userData.cameraInfo };
-  }
-}
 
 export class Images extends SceneExtension<ImageRenderable> {
   /* All known camera info topics */
@@ -242,44 +201,18 @@ export class Images extends SceneExtension<ImageRenderable> {
       this.cameraInfoToImageTopics.set(cameraInfoTopic, imageTopic);
     }
 
-    const cameraInfoTopicChanged =
-      prevCameraInfoTopic != undefined && cameraInfoTopic !== prevCameraInfoTopic;
-
-    // The topic did not change, some other setting changed
-    // We re-update the renderable with the new settings and are done
-    if (!cameraInfoTopicChanged) {
-      const renderable = this.renderables.get(imageTopic);
-      if (!renderable) {
-        return;
-      }
-
-      const { image, cameraInfo, receiveTime } = renderable.userData;
-      if (!image || !cameraInfo) {
-        return;
-      }
-
-      this._updateImageRenderable(renderable, image, cameraInfo, receiveTime, settings);
-      return;
-    }
-
-    // The camera info topic changed for our renderable
-    // Remove the previous camera_info_topic -> image_topic mapping
-    this.cameraInfoToImageTopics.delete(prevCameraInfoTopic, imageTopic);
-
     const renderable = this.renderables.get(imageTopic);
     if (!renderable) {
       return;
     }
 
-    // Create a new renderable preserving the image from the previous one
-    const { image, receiveTime, frameId } = renderable.userData;
+    renderable.setSettings(settings);
 
-    // We dispose the old renderable and setup a new one. It is easier to dispose the old
-    // renderable than clear only the specific camera topic fields.
-    renderable.dispose();
-    this.renderables.delete(imageTopic);
-
-    const newRenderable = this._getImageRenderable(imageTopic, receiveTime, image, frameId);
+    // The camera info topic changed for our renderable
+    // Remove the previous camera_info_topic -> image_topic mapping
+    if (prevCameraInfoTopic != undefined) {
+      this.cameraInfoToImageTopics.delete(prevCameraInfoTopic, imageTopic);
+    }
 
     // apply camera info to new renderable
     if (!cameraInfoTopic) {
@@ -296,10 +229,8 @@ export class Images extends SceneExtension<ImageRenderable> {
       );
       return;
     }
-
-    if (image) {
-      this._updateImageRenderable(newRenderable, image, cameraInfo, receiveTime, settings);
-    }
+    this._recomputeCameraModel(renderable, cameraInfo);
+    renderable.update();
   };
 
   private cameraInfoShouldSubscribe = (cameraInfoTopic: string): boolean => {
@@ -378,10 +309,6 @@ export class Images extends SceneExtension<ImageRenderable> {
         draft.topics[imageTopic] = updatedUserSettings;
       });
       this.updateSettingsTree();
-    }
-
-    // There's no camera info topic so we can't look up camera info messages to make a model
-    if (!settings.cameraInfoTopic) {
       return;
     }
 
@@ -396,7 +323,10 @@ export class Images extends SceneExtension<ImageRenderable> {
       return;
     }
 
-    this._updateImageRenderable(renderable, image, cameraInfo, receiveTime, settings);
+    this._recomputeCameraModel(renderable, cameraInfo);
+    renderable.userData.receiveTime = receiveTime;
+    renderable.setImage(image);
+    renderable.update();
   };
 
   private handleCameraInfo = (
@@ -422,16 +352,8 @@ export class Images extends SceneExtension<ImageRenderable> {
         continue;
       }
 
-      // If there's no active image then we don't need to update the renderable
-      //
-      // When an image is received, the handleImage call will look up the cameraInfoByTopic and
-      // create the model.
-      const { image, receiveTime } = renderable.userData;
-      if (!image) {
-        continue;
-      }
-
-      this._updateImageRenderable(renderable, image, cameraInfo, receiveTime, settings);
+      this._recomputeCameraModel(renderable, cameraInfo);
+      renderable.update();
     }
   };
 
@@ -443,159 +365,23 @@ export class Images extends SceneExtension<ImageRenderable> {
    *
    * This function will set a topic error on the image topic if the camera model creation fails.
    */
-  private _recomputeCameraModel(
-    renderable: ImageRenderable,
-    newCameraInfo: CameraInfo,
-  ): PinholeCameraModel | undefined {
+  private _recomputeCameraModel(renderable: ImageRenderable, newCameraInfo: CameraInfo) {
     // If the camera info has not changed, we don't need to make a new model and can return the existing one
     const dataEqual = cameraInfosEqual(renderable.userData.cameraInfo, newCameraInfo);
     if (dataEqual && renderable.userData.cameraModel != undefined) {
-      return renderable.userData.cameraModel;
+      return;
     }
 
     const imageTopic = renderable.userData.topic;
 
-    // clear the old model since that is no longer valid if the camera info changed
-    renderable.userData.cameraModel = undefined;
-    renderable.userData.geometry?.dispose();
-    renderable.userData.geometry = undefined;
-
     try {
-      renderable.userData.cameraModel = new PinholeCameraModel(newCameraInfo);
+      renderable.setCameraModel(new PinholeCameraModel(newCameraInfo));
+      renderable.userData.cameraInfo = newCameraInfo;
       this.renderer.settings.errors.removeFromTopic(imageTopic, CAMERA_MODEL);
     } catch (errUnk) {
       const err = errUnk as Error;
       this.renderer.settings.errors.addToTopic(imageTopic, CAMERA_MODEL, err.message);
-
-      // if there's no model, there's no way to update the renderable
-      return;
     }
-
-    // Save the latest camera info if we were able to make a model
-    // This is used to avoid recomputing the model if the camera info hasn't changed (above)
-    renderable.userData.cameraInfo = newCameraInfo;
-
-    return renderable.userData.cameraModel;
-  }
-
-  private _updateImageRenderable(
-    renderable: ImageRenderable,
-    image: AnyImage,
-    cameraInfo: CameraInfo,
-    receiveTime: bigint,
-    settings: Partial<LayerSettingsImage> | undefined,
-  ): void {
-    const prevSettings = renderable.userData.settings;
-    const newSettings = { ...DEFAULT_SETTINGS, ...settings };
-    const geometrySettingsEqual =
-      newSettings.cameraInfoTopic === prevSettings.cameraInfoTopic &&
-      newSettings.distance === prevSettings.distance &&
-      newSettings.planarProjectionFactor === prevSettings.planarProjectionFactor;
-    const materialSettingsEqual = newSettings.color === prevSettings.color;
-    const topic = renderable.userData.topic;
-
-    renderable.userData.image = image;
-    const cameraModel = (renderable.userData.cameraModel = this._recomputeCameraModel(
-      renderable,
-      cameraInfo,
-    ));
-
-    // We need a valid camera model to render the image
-    if (!cameraModel) {
-      return;
-    }
-
-    // If there is camera info, the frameId comes from the camera info since the user may have
-    // selected camera info with a different frame than our image frame.
-    //
-    // If there is no camera info, we fall back to the image's frame
-    const rawFrameId =
-      renderable.userData.cameraInfo?.header.frame_id ??
-      ("header" in image ? image.header.frame_id : image.frame_id);
-
-    renderable.userData.frameId = this.renderer.normalizeFrameId(rawFrameId);
-    renderable.userData.receiveTime = receiveTime;
-    renderable.userData.messageTime = toNanoSec(
-      "header" in image ? image.header.stamp : image.timestamp,
-    );
-    renderable.userData.settings = newSettings;
-
-    // Dispose of the current geometry if the settings have changed
-    if (!geometrySettingsEqual) {
-      renderable.userData.geometry?.dispose();
-      renderable.userData.geometry = undefined;
-      if (renderable.userData.mesh) {
-        renderable.remove(renderable.userData.mesh);
-        renderable.userData.mesh = undefined;
-      }
-    }
-
-    const hasCameraInfo = settings?.cameraInfoTopic != undefined;
-    if (hasCameraInfo) {
-      this.renderer.settings.errors.removeFromTopic(topic, NO_CAMERA_INFO_ERR);
-    }
-
-    // Create the plane geometry if needed
-    if (renderable.userData.geometry == undefined) {
-      const geometry = createGeometry(cameraModel, renderable.userData.settings);
-      renderable.userData.geometry = geometry;
-      if (renderable.userData.mesh) {
-        renderable.remove(renderable.userData.mesh);
-        renderable.userData.mesh = undefined;
-      }
-    }
-
-    // Create or update the bitmap texture
-    if ("format" in image) {
-      const bitmapData = new Blob([image.data], { type: `image/${image.format}` });
-      self
-        .createImageBitmap(bitmapData, { resizeWidth: DEFAULT_IMAGE_WIDTH })
-        .then((bitmap) => {
-          if (renderable.userData.texture == undefined) {
-            renderable.userData.texture = createCanvasTexture(bitmap);
-            rebuildMaterial(renderable);
-            tryCreateMesh(renderable, this.renderer);
-          } else {
-            renderable.userData.texture.image.close();
-            renderable.userData.texture.image = bitmap;
-            renderable.userData.texture.needsUpdate = true;
-          }
-
-          this.renderer.settings.errors.removeFromTopic(topic, CREATE_BITMAP_ERR);
-        })
-        .catch((err) => {
-          this.renderer.settings.errors.addToTopic(
-            topic,
-            CREATE_BITMAP_ERR,
-            `createBitmap failed: ${err.message}`,
-          );
-        });
-    } else {
-      const { width, height } = image;
-      const prevTexture = renderable.userData.texture as THREE.DataTexture | undefined;
-      if (
-        prevTexture == undefined ||
-        prevTexture.image.width !== width ||
-        prevTexture.image.height !== height
-      ) {
-        prevTexture?.dispose();
-        renderable.userData.texture = createDataTexture(width, height);
-        rebuildMaterial(renderable);
-        tryCreateMesh(renderable, this.renderer);
-      }
-
-      const texture = renderable.userData.texture as THREE.DataTexture;
-      rawImageToDataTexture(image, {}, texture);
-      texture.needsUpdate = true;
-    }
-
-    // Create or update the material if needed
-    if (!renderable.userData.material || !materialSettingsEqual) {
-      rebuildMaterial(renderable);
-    }
-
-    // Create/recreate the mesh if needed
-    tryCreateMesh(renderable, this.renderer);
   }
 
   // Get or create an image renderable for the imageTopic
@@ -638,156 +424,6 @@ export class Images extends SceneExtension<ImageRenderable> {
   }
 }
 
-type RawImageOptions = {
-  minValue?: number;
-  maxValue?: number;
-};
-
-const tempColor = { r: 0, g: 0, b: 0, a: 0 };
-
-function tryCreateMesh(renderable: ImageRenderable, renderer: IRenderer): void {
-  const { mesh, geometry, material } = renderable.userData;
-  if (!mesh && geometry && material) {
-    renderable.userData.mesh = new THREE.Mesh(geometry, renderable.userData.material);
-    renderable.add(renderable.userData.mesh);
-    renderer.queueAnimationFrame();
-  }
-}
-
-function rebuildMaterial(renderable: ImageRenderable): void {
-  const texture = renderable.userData.texture;
-
-  renderable.userData.material?.dispose();
-  renderable.userData.material = texture ? createMaterial(texture, renderable) : undefined;
-
-  // Destroy the mesh, it needs to be rebuilt
-  if (renderable.userData.mesh) {
-    renderable.remove(renderable.userData.mesh);
-    renderable.userData.mesh = undefined;
-  }
-}
-
-function createCanvasTexture(bitmap: ImageBitmap): THREE.CanvasTexture {
-  const texture = new THREE.CanvasTexture(
-    bitmap,
-    THREE.UVMapping,
-    THREE.ClampToEdgeWrapping,
-    THREE.ClampToEdgeWrapping,
-    THREE.NearestFilter,
-    THREE.LinearFilter,
-    THREE.RGBAFormat,
-    THREE.UnsignedByteType,
-  );
-  texture.generateMipmaps = false;
-  texture.encoding = THREE.sRGBEncoding;
-  return texture;
-}
-
-function createDataTexture(width: number, height: number): THREE.DataTexture {
-  const size = width * height;
-  const rgba = new Uint8ClampedArray(size * 4);
-  return new THREE.DataTexture(
-    rgba,
-    width,
-    height,
-    THREE.RGBAFormat,
-    THREE.UnsignedByteType,
-    THREE.UVMapping,
-    THREE.ClampToEdgeWrapping,
-    THREE.ClampToEdgeWrapping,
-    THREE.NearestFilter,
-    THREE.LinearFilter,
-    1,
-    THREE.sRGBEncoding,
-  );
-}
-
-function createMaterial(
-  texture: THREE.Texture,
-  renderable: ImageRenderable,
-): THREE.MeshBasicMaterial {
-  stringToRgba(tempColor, renderable.userData.settings.color);
-  const transparent = tempColor.a < 1;
-  const color = new THREE.Color(tempColor.r, tempColor.g, tempColor.b);
-  return new THREE.MeshBasicMaterial({
-    name: `${renderable.userData.topic}:Material`,
-    color,
-    map: texture,
-    side: THREE.DoubleSide,
-    opacity: tempColor.a,
-    transparent,
-    depthWrite: !transparent,
-  });
-}
-
-function createGeometry(
-  cameraModel: PinholeCameraModel,
-  settings: LayerSettingsImage,
-): THREE.PlaneGeometry {
-  const WIDTH_SEGMENTS = 10;
-  const HEIGHT_SEGMENTS = 10;
-
-  const width = cameraModel.width;
-  const height = cameraModel.height;
-  const geometry = new THREE.PlaneGeometry(1, 1, WIDTH_SEGMENTS, HEIGHT_SEGMENTS);
-
-  const gridX1 = WIDTH_SEGMENTS + 1;
-  const gridY1 = HEIGHT_SEGMENTS + 1;
-  const size = gridX1 * gridY1;
-
-  const segmentWidth = width / WIDTH_SEGMENTS;
-  const segmentHeight = height / HEIGHT_SEGMENTS;
-
-  // Use a slight offset to avoid z-fighting with the CameraInfo wireframe
-  const EPS = 1e-3;
-
-  // Rebuild the position buffer for the plane by iterating through the grid and
-  // proejcting each pixel space x/y coordinate into a 3D ray and casting out by
-  // the user-configured distance setting. UV coordinates are rebuilt so the
-  // image is not vertically flipped
-  const pixel = { x: 0, y: 0 };
-  const p = { x: 0, y: 0, z: 0 };
-  const vertices = new Float32Array(size * 3);
-  const uvs = new Float32Array(size * 2);
-  for (let iy = 0; iy < gridY1; iy++) {
-    for (let ix = 0; ix < gridX1; ix++) {
-      const vOffset = (iy * gridX1 + ix) * 3;
-      const uvOffset = (iy * gridX1 + ix) * 2;
-
-      pixel.x = ix * segmentWidth;
-      pixel.y = iy * segmentHeight;
-      projectPixel(p, pixel, cameraModel, settings);
-
-      vertices[vOffset + 0] = p.x;
-      vertices[vOffset + 1] = p.y;
-      vertices[vOffset + 2] = p.z - EPS;
-
-      uvs[uvOffset + 0] = ix / WIDTH_SEGMENTS;
-      uvs[uvOffset + 1] = iy / HEIGHT_SEGMENTS;
-    }
-  }
-
-  geometry.setAttribute("position", new THREE.BufferAttribute(vertices, 3));
-  geometry.setAttribute("uv", new THREE.BufferAttribute(uvs, 2));
-  geometry.attributes.position!.needsUpdate = true;
-  geometry.attributes.uv!.needsUpdate = true;
-
-  return geometry;
-}
-
-export function cameraInfoTopicMatches(imageTopic: string, cameraInfoTopic: string): boolean {
-  const imageParts = imageTopic.split("/");
-  const infoParts = cameraInfoTopic.split("/");
-
-  for (let i = 0; i < imageParts.length - 1 && i < infoParts.length - 1; i++) {
-    if (imageParts[i] !== infoParts[i]) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
 /**
  * Look up a matching camera info topic for the image topic.
  *
@@ -807,119 +443,15 @@ function autoSelectCameraInfoTopic(
   return candidates[0];
 }
 
-function rawImageToDataTexture(
-  image: RosImage | RawImage,
-  options: RawImageOptions,
-  output: THREE.DataTexture,
-): void {
-  const { encoding, width, height } = image;
-  const is_bigendian = "is_bigendian" in image ? image.is_bigendian : false;
-  const rawData = image.data as Uint8Array;
-  switch (encoding) {
-    case "yuv422":
-      decodeYUV(image.data as Int8Array, width, height, output.image.data);
-      break;
-    // same thing as yuv422, but a distinct decoding from yuv422 and yuyv
-    case "uyuv":
-      decodeYUV(image.data as Int8Array, width, height, output.image.data);
-      break;
-    // change name in the future
-    case "yuyv":
-      decodeYUYV(image.data as Int8Array, width, height, output.image.data);
-      break;
-    case "rgb8":
-      decodeRGB8(rawData, width, height, output.image.data);
-      break;
-    case "rgba8":
-      decodeRGBA8(rawData, width, height, output.image.data);
-      break;
-    case "bgra8":
-      decodeBGRA8(rawData, width, height, output.image.data);
-      break;
-    case "bgr8":
-    case "8UC3":
-      decodeBGR8(rawData, width, height, output.image.data);
-      break;
-    case "32FC1":
-      decodeFloat1c(rawData, width, height, is_bigendian, output.image.data);
-      break;
-    case "bayer_rggb8":
-      decodeBayerRGGB8(rawData, width, height, output.image.data);
-      break;
-    case "bayer_bggr8":
-      decodeBayerBGGR8(rawData, width, height, output.image.data);
-      break;
-    case "bayer_gbrg8":
-      decodeBayerGBRG8(rawData, width, height, output.image.data);
-      break;
-    case "bayer_grbg8":
-      decodeBayerGRBG8(rawData, width, height, output.image.data);
-      break;
-    case "mono8":
-    case "8UC1":
-      decodeMono8(rawData, width, height, output.image.data);
-      break;
-    case "mono16":
-    case "16UC1":
-      decodeMono16(rawData, width, height, is_bigendian, output.image.data, options);
-      break;
-    default:
-      throw new Error(`Unsupported encoding ${encoding}`);
+export function cameraInfoTopicMatches(imageTopic: string, cameraInfoTopic: string): boolean {
+  const imageParts = imageTopic.split("/");
+  const infoParts = cameraInfoTopic.split("/");
+
+  for (let i = 0; i < imageParts.length - 1 && i < infoParts.length - 1; i++) {
+    if (imageParts[i] !== infoParts[i]) {
+      return false;
+    }
   }
-}
 
-function normalizeImageData(data: Int8Array): Int8Array;
-function normalizeImageData(data: PartialMessage<Uint8Array> | undefined): Uint8Array;
-function normalizeImageData(data: unknown): Int8Array | Uint8Array;
-function normalizeImageData(data: unknown): Int8Array | Uint8Array {
-  if (data == undefined) {
-    return new Uint8Array(0);
-  } else if (data instanceof Int8Array || data instanceof Uint8Array) {
-    return data;
-  } else {
-    return new Uint8Array(0);
-  }
-}
-
-function normalizeRosImage(message: PartialMessage<RosImage>): RosImage {
-  return {
-    header: normalizeHeader(message.header),
-    height: message.height ?? 0,
-    width: message.width ?? 0,
-    encoding: message.encoding ?? "",
-    is_bigendian: message.is_bigendian ?? false,
-    step: message.step ?? 0,
-    data: normalizeImageData(message.data),
-  };
-}
-
-function normalizeRosCompressedImage(
-  message: PartialMessage<RosCompressedImage>,
-): RosCompressedImage {
-  return {
-    header: normalizeHeader(message.header),
-    format: message.format ?? "",
-    data: normalizeByteArray(message.data),
-  };
-}
-
-function normalizeRawImage(message: PartialMessage<RawImage>): RawImage {
-  return {
-    timestamp: normalizeTime(message.timestamp),
-    frame_id: message.frame_id ?? "",
-    height: message.height ?? 0,
-    width: message.width ?? 0,
-    encoding: message.encoding ?? "",
-    step: message.step ?? 0,
-    data: normalizeImageData(message.data),
-  };
-}
-
-function normalizeCompressedImage(message: PartialMessage<CompressedImage>): CompressedImage {
-  return {
-    timestamp: normalizeTime(message.timestamp),
-    frame_id: message.frame_id ?? "",
-    format: message.format ?? "",
-    data: normalizeByteArray(message.data),
-  };
+  return true;
 }
